requirements.txt
langchain-groq
langchain-community
langchain-core
langchain-text-splitters
sentence-transformers
faiss-cpu
pypdf
streamlit
torch
python-dotenv